Add scripts to source distribution using MANIFEST
The setup.py script is programmed to add either scripts/taskplot.cmd or
scripts/taskplot on Windows or Linux respectively. When source
distribution is created on Linux, only scripts/taskplot is included in
the source distribution. Therefore, installation of the program on
Windows fails due to missing scripts/taskplot.cmd. Hence, include both
scripts/taskplot.cmd and scripts/taskplot in the source distribution
irrespective of which system the source distribution to be released is
created on.